refactor(formatter): move `Expand` to `biome_formatter`
