+++ crates/biome_json_formatter/src/context.rs
@@ -2,7 +2,7 @@ use crate::JsonCommentStyle;
 use crate::comments::{FormatJsonLeadingComment, JsonComments};
 use biome_deserialize_macros::{Deserializable, Merge};
 use biome_formatter::separated::TrailingSeparator;
-use biome_formatter::{BracketSpacing, IndentWidth, ObjectWrap, prelude::*};
+use biome_formatter::{BracketSpacing, Expand, IndentWidth, ObjectWrap, prelude::*};
 use biome_formatter::{
     CstFormatContext, FormatContext, FormatOptions, IndentStyle, LineEnding, LineWidth,
     TransformSourceMap,
@@ -105,39 +105,6 @@ impl fmt::Display for TrailingCommas {
         match self {
             TrailingCommas::None => std::write!(f, "None"),
             TrailingCommas::All => std::write!(f, "All"),
-        }
-    }
-}
-
-#[derive(Clone, Copy, Debug, Default, Eq, Hash, Deserializable, Merge, PartialEq)]
-#[cfg_attr(
-    feature = "serde",
-    derive(serde::Serialize, serde::Deserialize),
-    serde(rename_all = "camelCase")
-)]
-#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
-pub enum Expand {
-    Always,
-    #[default]
-    FollowSource,
-}
-
-impl FromStr for Expand {
-    type Err = String;
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s {
-            "always" => Ok(Self::Always),
-            "follow-source" => Ok(Self::FollowSource),
-            _ => Err(std::format!("unknown expand literal: {}", s)),
-        }
-    }
-}
-
-impl fmt::Display for Expand {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            Expand::Always => std::write!(f, "Always"),
-            Expand::FollowSource => std::write!(f, "Follow Source"),
         }
     }
 }
